Add in pages for iOS app
Added in pages model in mongoose. Pages are used for dynamically creating views for the iOS app.
Pages should contain the image encoded in base64. mockdata-booklet.js contains the initial data.
This commit also adds in an API to view the Pages information in JSON or view the image in PNG.

Signed-off-by: AnthonyAmanse <ghieamanse@gmail.com>